profile fragment, cleanup UI
