+++ app/src/main/java/com/example/photofy/activities/MainActivity.java
@@ -20,6 +20,7 @@ import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.Volley;
 import com.example.photofy.R;
 import com.example.photofy.fragments.ComposeFragment;
+import com.example.photofy.fragments.ProfileFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 import com.parse.LogOutCallback;
@@ -31,7 +32,6 @@ public class MainActivity extends AppCompatActivity {
     public static final String TAG = "MainActivity";
 
     private BottomNavigationView bottomNavigationView;
-    private Button btnLogout;
     final FragmentManager fragmentManager = getSupportFragmentManager();
 
     private SharedPreferences.Editor editor;
@@ -50,7 +50,6 @@ public class MainActivity extends AppCompatActivity {
         editor.apply();
 
         bottomNavigationView = findViewById(R.id.bottomNavigation);
-        btnLogout = findViewById(R.id.btnLogout);
 
         bottomNavigationView.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
             @Override
@@ -60,11 +59,13 @@ public class MainActivity extends AppCompatActivity {
                     case R.id.action_home:
                         // TODO: update fragment
                         fragment = new ComposeFragment();
+                        break;
                     case R.id.action_compose:
                         fragment = new ComposeFragment();
+                        break;
                     case R.id.action_profile:
-                        // TODO: update fragment
-                        fragment = new ComposeFragment();
+                        fragment = new ProfileFragment();
+                        break;
                     default:
                         break;
                 }
@@ -74,30 +75,6 @@ public class MainActivity extends AppCompatActivity {
         });
         // Set default selection
         bottomNavigationView.setSelectedItemId(R.id.action_home);
-
-        btnLogout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                logout();
-            }
-        });
-    }
-
-    // TODO: use AuthorizationClient clearCookies method to log out and clear all stored tokens for Spotify
-    private void logout() {
-        ParseUser.logOutInBackground(new LogOutCallback() {
-            @Override
-            public void done(ParseException e) {
-                if (e != null) {
-                    Toast.makeText(MainActivity.this, R.string.logout_error_toast, Toast.LENGTH_SHORT).show();
-                    Log.e(TAG, "Issue with logout", e);
-                } else {
-                    goLoginActivity();
-                    Toast.makeText(MainActivity.this, R.string.logout_toast, Toast.LENGTH_SHORT).show();
-                }
-            }
-        });
-        ParseUser currentUser = ParseUser.getCurrentUser();
     }
 
     private void goLoginActivity() {
