+++ app/src/main/res/layout/activity_main.xml
@@ -16,12 +16,6 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" >
-
-        <Button
-            android:id="@+id/btnLogout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="LOGOUT" />
     </FrameLayout>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
