+++ app/src/main/res/layout/fragment_compose.xml
@@ -7,29 +7,48 @@
     android:layout_height="match_parent"
     tools:context=".fragments.ComposeFragment">
 
-    <Button
-        android:id="@+id/btnEnableCamera"
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/innerconstraint"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/open_camera"
+        android:layout_marginTop="15dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent">
+
+        <Button
+            android:id="@+id/btnEnableCamera"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="15dp"
+            android:text="@string/open_camera"
+            app:layout_constraintEnd_toStartOf="@+id/button"
+            app:layout_constraintStart_toStartOf="parent"
+            tools:layout_editor_absoluteY="0dp" />
+
+        <Button
+            android:id="@+id/button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Upload image"
+            app:layout_constraintEnd_toEndOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <ImageView
         android:id="@+id/ivCapturedImage"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:contentDescription="@string/captured_image"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/btnGetColors"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/btnEnableCamera" />
+        app:layout_constraintTop_toBottomOf="@+id/innerconstraint" />
 
     <Button
         android:id="@+id/btnGetColors"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginBottom="15dp"
         android:text="Get colors"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
