+++ app/src/main/res/values/strings.xml
@@ -14,4 +14,6 @@
     <string name="sign_up_here">Sign up here</string>
     <string name="new_account">Don\'t have an account?</string>
     <string name="photofy">PHOTOFY</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
